Association of Equipment Providers Cooperative
The Association of Equipment Providers Cooperative (AEPC) is a database-driven system designed to manage equipment orders, member information, messaging, and related data for a cooperative of equipment providers. The system streamlines the process of placing orders, tracking inventory, and facilitating communication between customers and cooperative members. Project programmed with SQL.